Change conserve to a bin-only crate
Unblocks using `cargo check` and clippy.

This also gives warnings about code never used by the tool.

Unfortunately it breaks doctests, which will need to be converted.

This also requires a stricter separation of the blackbox tests, which
can no longer link to conserve code, but that may be good.