Extract StoreFiles object
This is a layer through which files are
written into the BlockDir.

In this version, it reuses an input buffer per file that's stored, avoiding a 1MB allocation for every file.

In future I hope this can be used to
block, by returning a promise of a future address. Or, to compress files in parallel, on a thread pool.