Speed up LiveTree::iter
This function is required to produce entries in Apath order, which is
not a naive string order.  Previously it achieved this by comparing the
Apath (tree-relative path) for newly generated entries.

But entries are generated one directory at a time, so their paths have a
long common prefix. Profiles show it spent a long time walking and
comparing the string prefixes, and this is unnecessary.

This change should speed up tree listing and therefore backups where
few files have changed.