Remove check on blocks of unchanged files
This cuts the time to backup a large tree with few changes to a
moderately-slow USB flash drive by ~7x.

Although it seemed prudent to check the blocks are there, it's not
actually necessary unless the archive has previously lost data.
Blocks are always written before the index that points to them, so if
the index hunk is there, we can safely assume the blocks are there.

Rather than ad-hoc expensive checks for invariants during backup and
restore, it's probably better to make them fast on their preconditions,
and check the invariants from `validate`.