Convert Archive to use Transport
This should remove all remaining assumptions that the archive is on a
local filesystem, although no other transports are implemented yet.

This also allows deleting some now unused code in jsonio and io. In
particular, AtomicFile moves to being effectively handled in the
Transport, because some may be able to atomically write a file.

Fixes #13.