perf: Read blocks only once during validate
Previously, blocks were read once while validating the blockdir, then
again for each file that references them.

With this change, blocks are each read once when checking the blockdir,
and we remember the uncompressed length. Then to check indexes, we only
need to check that all addresses point within the uncompressed range of
blocks that are present.