Use BlockHash type everywhere, rather than strings
This should use less memory, for cases such as gc or validation where
there are many hashes in memory simultaneously. It's more compact and
also stored inline rather than a String on the heap.

A downside is that it needs to be converted to and from hex strings for
filenames or serialization.