validate need not lstat all blocks
On macOS (at least), the most time-consuming part of initial validation
is lstating the files. It's not really necessary: we can just list the
directories to see which blocks are present.

This does remove the progress bar for validation, but it can be added
back later, and the performance win is worth it.

To do this, the transport DirEntry no longer includes the length, but a
new Metadata type is added, analogous to std::fs.

Fixes #128