Mark archive locked during GC
For safety, we need to prevent new backups starting while a gc operation
is underway, otherwise they might reference blocks that will soon be
removed.

This adds a GC_LOCK file in the archive directory which is present only
while GC is underway.

Only one GC can run at a time.

The lock is removed on drop, and can be explictly broken.

This is not yet wired into backups or the CLI.