Sort entries just before writing index hunks
Previously the API required they be added to the IndexBuilder in order,
but that gets in the way of grouping together small files and
parallelizing compression.

Instead, allow them to be added in arbitrary order within the hunk, and
the IndexWriter will sort before serializing.  Backup code is still
responsible for keeping them in order across hunks, and the IndexWriter
checks this.

Also, rename IndexBuilder to IndexWriter.