Add --log-json, remove --problems-json
Instead of writing a structured log of errors in the Monitor,
hook this into Tracing. That means these can be directly
emitted with error!() like macros which should be simpler, and
will get the right code location.

Also instead of counting errors and warnings in the monitor,
count them in a custom tracing Layer.